Adaptive smoothing by Intersection of Confidence Intervals (ICI) using multiple windows

The function performs adaptive smoothing by Intersection of Confidence Intervals (ICI) using multiple windows as described in Katkovnik et al (2006)

ICIcombined(y, hmax, hinc = 1.45, thresh = NULL, kern = "Gaussian", m = 0,
            sigma = NULL, nsector = 1, symmetric = FALSE, presmooth = FALSE,
            combine = "weighted", unit = c("SD","FWHM"))

Arguments

y
Object of class "array" containing the original (response) data on a grid

hmax
maximum bandwidth

hinc
factor used to increase the bandwidth from scale to scale

thresh
threshold used in tests to determine the best scale

kern
Determines the kernel function. Object of class "character" kernel, can be any of c("Gaussian","Uniform","Triangle","Epanechnicov","Biweight","Triweight"). Defaults to kern="Gaussian".

m
Object of class "integer" vector of length length(dy) determining the order of derivatives specified for the coordinate directios.

sigma
error standard deviation

nsector
number of sectors to use.

symmetric
Object of class "logical" determines if sectors are symmetric with respect to the origin.

presmooth
Object of class "logical" determines if bandwidths are smoothed for more stable results.

combine
Either "weighted" or "minvar". Determines how whether to combine sectorial results a weighted (with inverse variance) mean or to chose the sectorial estimate with minimal variance.

unit
How should the bandwidth be interpreted in case of a Gaussian kernel. For "SD" the bandwidth refers to the standard deviation of the kernel while "FWHM" interprets the banwidth in terms of Full Width Half Maximum of the kernel.

Details

This mainly follows Chapter 6.2 in Katkovnik et al (2006).

Value

An object of class ICIsmooth

References

V. Katkovnik, K. Egiazarian and J. Astola, Local Approximation Techniques in Signal And Image Processing, SPIE Society of Photo-Optical Instrumentation Engin., 2006, PM157

Author

Joerg Polzehl polzehl@wias-berlin.de

See also