Class "ICIsmooth"

The "ICIsmooth" class is used for objects obtained by functions ICIsmooth and ICIcombined.

Objects from the Class

Objects can be created by calls of the form new("ICIsmooth", ...) or by functions ICIsmooth and ICIcombined.

Slots

.Data:

Object of class "list", usually empty.

y:

Object of class "array" containing the original (response) data

dy:

Object of class "numeric" dimension attribute of y

x:

Object of class "numeric" if provided the design points

hmax:

Object of class "numeric" maximum bandwidth

hinc:

Object of class "numeric" initial bandwidth

thresh:

Object of class "numeric" threshold used for bandwidth selection

kern:

Object of class "character" kernel, can be any of c("Gaussian","Uniform","Triangle","Epanechnicov","Biweight","Triweight"). Defaults to kern="Gaussian".

m:

Object of class "integer" vector of length length(dy) determining the order of derivatives specified for the coordinate directios.

nsector:

Object of class "integer" number of sectors to use.

sector:

Object of class "integer" sector used.

symmetric:

Object of class "logical" sectors are symmetric with respect to the origin.

yhat:

Object of class "array" smoothed response variable

vhat:

Object of class "array" estimated variance of smoothed response variable

hbest:

Object of class "array" selected bandwidth(s))

sigma:

Object of class "numeric" estimated standard deviation of errors in y

call:

Object of class "call" that created the object.

Methods

extract

signature(x = "ICIsmooth"): ...

risk

signature(y = "ICIsmooth"): ...

plot

Method for Function `plot' in Package `aws'.

show

Method for Function `show' in Package `aws'.

print

Method for Function `print' in Package `aws'.

summary

Method for Function `summary' in Package `aws'.

References

V. Katkovnik, K. Egiazarian and J. Astola, Local Approximation Techniques in Signal And Image Processing, SPIE Society of Photo-Optical Instrumentation Engin., 2006, PM157

Author

Joerg Polzehl polzehl@wias-berlin.de

See also

Examples

showClass("ICIsmooth")
#> Class "ICIsmooth" [package "aws"]
#> 
#> Slots:
#>                                                                             
#> Name:      .Data         y        dy         x      hmax      hinc    thresh
#> Class:      list     array   numeric   numeric   numeric   numeric   numeric
#>                                                                             
#> Name:       kern         m   nsector    sector symmetric      yhat      vhat
#> Class: character   integer   integer   integer   logical     array     array
#>                                     
#> Name:      hbest     sigma      call
#> Class:     array   numeric      call