Class "aws"

The "aws" class is used for objects obtained by functions aws, lpaws, aws.irreg and aws.gaussian.

Objects from the Class

Objects are created by calls to functions aws, lpaws, aws.irreg and aws.gaussian.

Slots

.Data:

Object of class "list", usually empty.

y:

Object of class "array" containing the original (response) data

dy:

Object of class "numeric" dimension attribute of y

nvec:

Object of class "integer" leading dimension of y in vector valued data.

x:

Object of class "numeric" if provided the design points

ni:

Object of class "numeric" sum of weights used in final estimate

mask:

Object of class "logical" mask of design points where computations are performed

theta:

Object of class "array" containes the smoothed object and in case of function lpaws its derivatives up to the specified degree. Dimension is dim(theta)=c(dy,p)

hseq:

Sequence of bandwidths employed.

mae:

Object of class "numeric" Mean absolute error with respect to array in argument u if provided.

psnr:

Object of class "numeric" Peak Signal to Noise Ratio (PSNR) with respect to array in argument u if provided.

var:

Object of class "numeric" pointwise variance of theta[...,1]

xmin:

Object of class "numeric" min of x in case of irregular design

xmax:

Object of class "numeric" max of x in case of irregular design

wghts:

Object of class "numeric" weights used in location penalty for different coordinate directions, corresponds to ratios of distances in coordinate directions 2 and 3 to and distance in coordinate direction 1.

degree:

Object of class "integer" degree of local polynomials used in function lpaws

hmax:

Object of class "numeric" maximal bandwidth

sigma2:

Object of class "numeric" estimated error variance

scorr:

Object of class "numeric" estimated spatial correlation

family:

Object of class "character" distribution of y, can be any of c("Gaussian","Bernoulli","Poisson","Exponential", "Volatility","Variance")

shape:

Object of class "numeric" possible shape parameter of distribution of y

lkern:

Object of class "integer" location kernel, can be any of c("Triangle","Quadratic","Cubic","Plateau","Gaussian"), defauts to "Triangle"

lambda:

Object of class "numeric" scale parameter used in adaptation

ladjust:

Object of class "numeric" factor to adjust scale parameter with respect to its predetermined default.

aws:

Object of class "logical" Adaptation by Propagation-Separation

memory:

Object of class "logical" Adaptation by Stagewise Aggregation

homogen:

Object of class "logical" detect regions of homogeneity (used to speed up the calculations)

earlystop:

Object of class "logical" further speedup in function lpaws estimates are fixed if sum of weigths does not increase with iterations.

varmodel:

Object of class "character" variance model used in function aws.gaussian

vcoef:

Object of class "numeric" estimates variance parameters in function aws.gaussian

call:

Object of class "call" that created the object.

Methods

extract

signature(x = "aws"): ...

risk

signature(y = "aws"): ...

plot

Method for Function `plot' in Package `aws'.

show

Method for Function `show' in Package `aws'.

print

Method for Function `print' in Package `aws'.

summary

Method for Function `summary' in Package `aws'.

References

Joerg Polzehl, Vladimir Spokoiny, Adaptive Weights Smoothing with applications to image restoration, J. R. Stat. Soc. Ser. B Stat. Methodol. 62 , (2000) , pp. 335--354

Joerg Polzehl, Vladimir Spokoiny, Propagation-separation approach for local likelihood estimation, Probab. Theory Related Fields 135 (3), (2006) , pp. 335--362.

Author

Joerg Polzehl, polzehl@wias-berlin.de

See also

Examples

showClass("aws")
#> Class "aws" [package "aws"]
#> 
#> Slots:
#>                                                                             
#> Name:      .Data         y        dy      nvec         x        ni      mask
#> Class:      list     array   numeric   integer    matrix     array   logical
#>                                                                             
#> Name:      theta       mae      psnr       var      xmin      xmax     wghts
#> Class:     array   numeric   numeric   numeric   numeric   numeric   numeric
#>                                                                             
#> Name:     degree      hmax      hseq    sigma2     scorr    family     shape
#> Class:   integer   numeric   numeric   numeric   numeric character   numeric
#>                                                                             
#> Name:      lkern    lambda   ladjust       aws    memory   homogen earlystop
#> Class:   integer   numeric   numeric   logical   logical   logical   logical
#>                                     
#> Name:   varmodel     vcoef      call
#> Class: character   numeric      call