Class "awssegment"

The "aws" class is used for objects obtained by functions aws.segment

Objects from the Class

Objects are created by calls to functions aws.segment

Slots

.Data:

Object of class "list", usually empty.

y:

Object of class "array" containing the original (response) data

dy:

Object of class "numeric" dimension attribute of y

x:

Object of class "numeric" if provided the design points

ni:

Object of class "numeric" sum of weights used in final estimate

mask:

Object of class "logical" mask of design points where computations are performed

segment:

Object of class "array" segmentation results (3 segments coded by c(-1, 0, 1))

level:

Object of class "numeric" center of segment 0

delta:

Object of class "numeric" half width of segment 0

theta:

Object of class "array" ~~

theta:

Object of class "array" containes the smoothed object and in case of function lpaws its derivatives up to the specified degree. Dimension is dim(theta)=c(dy,p)

mae:

Object of class "numeric" Mean absolute error with respect to array in argument u if provided.

var:

Object of class "numeric" pointwise variance of theta[...,1]

xmin:

Object of class "numeric" not used

xmax:

Object of class "numeric" not used

wghts:

Object of class "numeric" weights used in location penalty for different coordinate directions

degree:

not used

hmax:

Object of class "numeric" maximal bandwidth

sigma2:

Object of class "numeric" estimated error variance

scorr:

Object of class "numeric" estimated spatial correlation

family:

Object of class "character" distribution of y, can be any of c("Gaussian","Bernoulli","Poisson","Exponential", "Volatility","Variance")

shape:

Object of class "numeric" possible shape parameter of distribution of y

lkern:

Object of class "integer" location kernel, can be any of c("Triangle","Quadratic","Cubic","Plateau","Gaussian"), defauts to "Triangle"

lambda:

Object of class "numeric" scale parameter used in adaptation

ladjust:

Object of class "numeric" factor to adjust scale parameter with respect to its predetermined default.

aws:

Object of class "logical" Adaptation by Propagation-Separation

memory:

Object of class "logical" Adaptation by Stagewise Aggregation

homogen:

Object of class "logical" detect regions of homogeneity (used to speed up the calculations) currently FALSE

earlystop:

Object of class "logical" currently FALSE

varmodel:

Object of class "character" variance model used currently "Gaussian"

vcoef:

Object of class "numeric" contains NULL

call:

Object of class "call" that created the object.

Methods

extract

signature(x = "awssegment"): ...

plot

signature(x = "awssegment"): ...

print

signature(x = "awssegment"): ...

risk

signature(y = "awssegment"): ...

show

signature(object = "awssegment"): ...

summary

signature(object = "awssegment"): ...

Author

Joerg Polzehl, polzehl@wias-berlin.de

See also

Examples

showClass("awssegment")
#> Class "awssegment" [package "aws"]
#> 
#> Slots:
#>                                                                             
#> Name:      .Data         y        dy         x        ni      mask   segment
#> Class:      list     array   numeric   numeric     array   logical     array
#>                                                                             
#> Name:      level     delta     theta       mae       var      xmin      xmax
#> Class:   numeric   numeric     array   numeric   numeric   numeric   numeric
#>                                                                             
#> Name:      wghts    degree      hmax    sigma2     scorr    family     shape
#> Class:   numeric   integer   numeric   numeric   numeric character   numeric
#>                                                                             
#> Name:      lkern    lambda   ladjust       aws    memory   homogen earlystop
#> Class:   integer   numeric   numeric   logical   logical   logical   logical
#>                                     
#> Name:   varmodel     vcoef      call
#> Class: character   numeric      call