Propagation condition for adaptive weights smoothing

The function enables testing of the propagation condition in order to select appropriate values for the parameter lambda in function aws.

awstestprop(dy, hmax, theta = 1, family = "Gaussian", lkern = "Triangle",
            aws = TRUE, memory = FALSE, shape = 2, homogeneous=TRUE, varadapt=FALSE,
            ladjust = 1, spmin=0.25, seed = 1, minlevel=1e-6, maxz=25, diffz=.5,
            maxni=FALSE, verbose=FALSE)
pawstestprop(dy, hmax, theta = 1, family = "Gaussian", lkern = "Triangle",
             aws = TRUE, patchsize=1, shape = 2,
             ladjust = 1, spmin = 0.25, seed = 1, minlevel = 1e-6,
             maxz = 25, diffz = .5, maxni = FALSE, verbose = FALSE)

Arguments

dy
Dimension of grid used in 1D, 2D or 3D. May also be specified as an array of values. In this case data are generated with parameters dy-mean(dy)+theta and the propagation condition is testet as if theta is the true parameter. This can be used to study properties for a slighty misspecified structural assumption.

hmax
Maximum bandwidth.

theta
Parameter determining the distribution in case of family %in% c("Poisson","Bernoulli")

family
family specifies the probability distribution. Default is family="Gaussian", also implemented are "Bernoulli", "Poisson", "Exponential", "Volatility", "Variance" and "NCchi". family="Volatility" specifies a Gaussian distribution with expectation 0 and unknown variance. family="Volatility" specifies that p*y/theta is distributed as \(\chi^2\) with p=shape degrees of freedom. family="NCchi" uses a noncentral Chi distribution with p=shape degrees of freedom and noncentrality parameter theta.

lkern
character: location kernel, either "Triangle", "Plateau", "Quadratic", "Cubic" or "Gaussian"

aws
logical: if TRUE structural adaptation (AWS) is used.

patchsize
patchsize in case of paws.

memory
logical: if TRUE stagewise aggregation is used as an additional adaptation scheme.

shape
Allows to specify an additional shape parameter for certain family models. Currently only used for family="Variance", that is \(\chi\)-Square distributed observations with shape degrees of freedom.

homogeneous
if homgeneous==FALSE and family==Gaussian then create heterogeneous variances according to a chi-squared distribution with number of degrees of freedom given by sphere

varadapt
if varadapt==TRUE use inverse of variance reduction instead of sum of weights in definition of statistical penalty.

ladjust
Factor to increase the default value of lambda

spmin
Determines the form (size of the plateau) in the adaptation kernel. Not to be changed by the user.

seed
Seed value for random generator.

minlevel
Minimum exceedence probability to use in contour plots.

maxz
Maximum of z-scale in plots.

diffz
Gridlength in z

maxni
If TRUE use \(max_{l<=k}(N_i^{(l)}\) instead of \((N_i^{(k)}\) in the definition of the statistical penalty.

verbose
If TRUE provide additional information.

Details

Estimates exceedence probabilities

Results for intermediate steps are provided as contour plots. For a good choice of lambda (ladjust) the contours up to probabilities of 1e-5 should be vertical.

Value

A list with components

h

Sequence of bandwidths used

z

seq(0,30,.5), the quantiles exceedence probabilities refer to

prob

the matrix of exceedence probabilities, columns corresponding to h

probna

the matrix of exceedence probabilities for corresponding nonadaptive estimates, columns corresponding to h

References

Becker (2013)

Author

Joerg Polzehl polzehl@wias-berlin.de

See also