Methods for Function extract in Package aws

The method extract and/or compute specified statistics from object of class "aws", "awssegment", ICIsmooth and "kernsm".

# S4 method for aws
extract(x, what="y")
  # S4 method for awssegment
extract(x, what="y")
  # S4 method for ICIsmooth
extract(x, what="y")
  # S4 method for kernsm
extract(x, what="y")

Arguments

x
object

what
Statistics to extract, defaults to what="y" corresponding to the original data (response variable). Alternatives are what="yhat" for the smoothed response, what="vhat" for the estimated variance of the smoothed response, what="sigma2" for the estimated error variance of the original data, what="vred" for the variance reduction achieved and in case of signature(x = "ICIsmooth") what="hbest" for the selected bandwidth. A vector of any of these choices may be provided.

Methods

signature(x = "ANY")

Returns a message that method extract is not defined.

signature(x = "aws")

Returns a list with components containing the requested statistics. Component names correspond to tolower(what)

signature(x = "awssegment")

Returns a list with components containing the requested statistics. Component names correspond to tolower(what)

signature(x = "ICIsmooth")

Returns a list with components containing the requested statistics. Component names correspond to tolower(what).

signature(x = "kernsm")

Returns a list with components containing the requested statistics. Component names correspond to tolower(what).