Class "kernsm"

This class refers to objects created by function kernsm. These objects contain

Objects from the Class

Objects can be created by calls of the form new("kernsm", ...). they are usually created by a call to function{kernsm}.

Slots

.Data:

Object of class "list", usually empty.

y:

Object of class "array" containing the response in nonparametric regression. The design is assumed to be a 1D, 2D or 3D grid, with dimensionality determined by dim(y).

dy:

Object of class "numeric" containing dim(y).

x:

Object of class "numeric" currently not used.

h:

Object of class "numeric" containing the bandwidth employed.

kern:

Object of class "character" determining the kernel that was used, can be one of c("Gaussian","Uniform","Triangle","Epanechnikov","Biweight","Triweight")

m:

Object of class "integer" with length length(dy) determining the order of derivatives in the corresponding coordinate directions. If m[i6>0] a dirivative kernel derived from kern has been used for the corresponding coordinate direction.

nsector:

Object of class "integer". If nsector>1 positive weights are restricted to a segment of a circle (1D or 2D only). The segment is given by sector.

sector:

Object of class "integer" containing the number of the segment used in case of nsector>1

symmetric:

Object of class "logical" determines if the sector is mirrored at the origin.

yhat:

Object of class "array" with same size and dimension as y providing the convolution of y with the chosen kernel.

vred:

Object of class "array" Variance reduction achieved by convolution assuming independence.

call:

Object of class "function", call that created the object.

Methods

extract

signature(x = "aws"): ...

risk

signature(y = "aws"): ...

plot

Method for Function `plot' in Package `aws'.

show

Method for Function `show' in Package `aws'.

print

Method for Function `print' in Package `aws'.

summary

Method for Function `summary' in Package `aws'.

Author

J\"org Polzehl polzehl@wias-berlin.de

See also

Examples

showClass("kernsm")
#> Class "kernsm" [package "aws"]
#> 
#> Slots:
#>                                                                             
#> Name:      .Data         y        dy         x         h      kern         m
#> Class:      list     array   numeric   numeric   numeric character   integer
#>                                                                   
#> Name:    nsector    sector symmetric      yhat      vred      call
#> Class:   integer   integer   logical     array   numeric      call