Adaptive smoothing in orientation space SE(3)

The functions perform adaptive weights smoothing for data in orientation space SE(3), e.g. diffusion weighted MR data, with spatial coordinates given by voxel location within a mask and spherical information given by gradient direction. Observations can belong to different shells characterized by b-value bv. The data provided should only refer to voxel within mask.

smse3ms(sb, s0, bv, grad, kstar, lambda, kappa0, mask, sigma,
    ns0 = 1, ws0 = 1, vext = NULL, ncoils = 1, verbose = FALSE, usemaxni = TRUE)
smse3(sb, s0, bv, grad, mask, sigma, kstar, lambda, kappa0,
    ns0 = 1, vext = NULL, vred = 4, ncoils = 1, model = 0, dist = 1,
    verbose = FALSE)

Arguments

sb
2D array of diffion weighted data, first dimension refers to index ov voxel within the mask, second dimension to the number diffusion weighted images.

s0
vector of length sum(mask) containing values within mask of an average non-diffusion-weigthed image.

bv
vector of b-values.

grad
matrix of gradient directions with dim(grad)[1]==3.

kstar
number of steps in adaptive weights smoothing.

lambda
Scale parameter in adaptation

kappa0
determines amount of smoothing on the sphere. Larger values correspond to stronger smoothing on the sphere. If kappa0=NULL a value is that corresponds to a variace reduction with factor vred on the sphere.

mask
3D image defining a mask (logical)

sigma
Error standard deviation. Assumed to be known and homogeneous in the current implementation. A reasonable estimate may be defined as the modal value of standard deviations obtained using method getsdofsb.

ns0
Actual number of non-diffusion-weigthed images used to obtain s0 by averaging.

ws0
Weight for non-diffusion-weigthed images in statistical penalty.

vext
Voxel extensions.

ncoils
Effective number of receiver coils (in case of e.g. GRAPPA reconstructions), should be 1 in case of SENSE reconstructions. 2*ncoils is the number of degrees of freedom of the intensity distribution used.

verbose
If verbose=TRUE additional reports are given.

usemaxni
If "usemaxni==TRUE" a strikter penalization is used.

vred
Used if kappa0=NULL to specify the variance reduction on the sphere when suggesting a value of kappa0.

model
Determines which quantities are smoothed. Possible values are "Chi" for observed values (assumed to be distributed as noncentral Chi with 2*ncoils degrees of freedom), "Chi2" for squares of observed values (assumed to be distributed as noncentral Chi-squared with 2*ncoils degrees of freedom). "Gapprox" and "Gapprox2" use a Gaussian approximation for the noncentral Chi distribution to smooth ovserved and squared values, respectively.

dist
Distance in SE3. Reasonable values are 1 (default, see Becker et.al. 2012), 2 ( a slight modification of 1: with k6^2 instead of abs(k6)) and 3 (using a 'naive' distance on the sphere)

Value

The functions return lists with main results in components th and th0 containing the smoothed data.

References

Joerg Polzehl, Karsten Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Doi:10.1007/978-3-030-29184-6.

S. Becker, K. Tabelow, H.U. Voss, A. Anwander, R. Heidemann, J. Polzehl. Position-orientation adaptive smoothing of diffusion weighted magnetic resonance data (POAS). Medical Image Analysis, 2012, 16, 1142-1155. DOI:10.1016/j.media.2012.05.007.

S. Becker, K. Tabelow, S. Mohammadi, N. Weiskopf, J. Polzehl. Adaptive smoothing of multi-shell diffusion-weighted magnetic resonance data by msPOAS. Neuroimage, 2014, 95, 90-105. DOI:10.1016/j.neuroimage.2014.03.053.

Author

J\"org Polzehl polzehl@wias-berlin.de

Note

These functions are intended to be used internally in package dti only.