Methods for Function `summary' from package 'base' in Package `aws'

The method provides summary information for objects of class "aws".

Arguments

object
Object of class "dti", "dtiData", "dtiTensor", "dwiMixtensor", "dtiIndices", "dwiQball" or "dwiFiber".

...
Additional arguments in ... are passed to function quantile, e.g. argument probs may be specified here.

Methods

signature(object = "ANY")

Generic function: see summary.

signature(object = "aws")

The function provides summary information for objects of class "aws"

signature(object = "awssegment")

The function provides summary information for objects of class "awssegment"

signature(object = "ICIsmooth")

The function provides summary information for objects of class "ICIsmooth"

signature(object = "kernsm")

The function provides summary information for objects of class "kernsm"

Author

J\"org Polzehl polzehl@wias-berlin.de

See also