vector valued version of function aws The function implements the propagation separation approach to nonparametric smoothing (formerly introduced as Adaptive weights smoothing) for varying coefficient likelihood models with vector valued response on a 1D, 2D or 3D grid.

The function implements a version the propagation separation approach that uses vector valued instead of scalar responses.

vaws(y, kstar = 16, sigma2 = 1, mask = NULL, scorr = 0, spmin = 0.25,
     ladjust = 1, wghts = NULL, u = NULL, maxni = FALSE)
vawscov(y, kstar = 16, invcov = NULL, mask = NULL, scorr = 0, spmin = 0.25,
          ladjust = 1, wghts = NULL, u = NULL, maxni = FALSE)

Arguments

y
y contains the observed response data. dim(y) determines the dimensionality and extend of the grid design. First component varies over components of the response vector.

kstar
maximal number of steps to employ. Determines maximal bandwidth.

sigma2
specifies a homogeneous error variance.

invcov
array of voxelwise inverse covariance matrixes, first index corresponds to upper diagonal inverse covariance matrix.

mask
logical mask. All computations are restrikted to design poins within the mask.

scorr
The vector scorr allows to specify a first order correlations of the noise for each coordinate direction, defaults to 0 (no correlation).

spmin
determines the form (size of the plateau) in the adaptation kernel. Not to be changed by the user.

ladjust
factor to increase the default value of lambda

wghts
wghts specifies the diagonal elements of a weight matrix to adjust for different distances between grid-points in different coordinate directions, i.e. allows to define a more appropriate metric in the design space.

u
a "true" value of the regression function, may be provided to report risks at each iteration. This can be used to test the propagation condition with u=0

maxni
If TRUE use \(max_{l<=k}(N_i^{(l)}\) instead of \((N_i^{(k)}\) in the definition of the statistical penalty.

Details

see aws. Expets vector valued responses. Currently only implements the case of additive Gaussian errors.

Value

returns anobject of class aws with slots

y = "numeric"

y

dy = "numeric"

dim(y)

x = "numeric"

numeric(0)

ni = "integer"

integer(0)

mask = "logical"

logical(0)

theta = "numeric"

Estimates of regression function, length: length(y)

hseq = "numeric"

sequence of bandwidths employed

mae = "numeric"

Mean absolute error for each iteration step if u was specified, numeric(0) else

psnr = "numeric"

Peak signal-to-noise ratio for each iteration step if u was specified, numeric(0) else

var = "numeric"

approx. variance of the estimates of the regression function. Please note that this does not reflect variability due to randomness of weights.

xmin = "numeric"

numeric(0)

xmax = "numeric"

numeric(0)

wghts = "numeric"

numeric(0), ratio of distances wghts[-1]/wghts[1]

degree = "integer"

0

hmax = "numeric"

effective hmax

sigma2 = "numeric"

provided or estimated (inverse) error variance

scorr = "numeric"

scorr

family = "character"

family

shape = "numeric"

shape

lkern = "integer"

integer code for lkern, 1="Plateau", 2="Triangle", 3="Quadratic", 4="Cubic", 5="Gaussian"

lambda = "numeric"

effective value of lambda

ladjust = "numeric"

effective value of ladjust

aws = "logical"

aws

memory = "logical"

memory

homogen = "logical"

homogen

earlystop = "logical"

FALSE

varmodel = "character"

"Constant"

vcoef = "numeric"

numeric(0)

call = "function"

the arguments of the call to aws

References

J. Polzehl, K. Tabelow (2019). Magnetic Resonance Brain Imaging: Modeling and Data Analysis Using R. Springer, Use R! series. Appendix A. Doi:10.1007/978-3-030-29184-6.

J. Polzehl, V. Spokoiny, Adaptive Weights Smoothing with applications to image restoration, J. R. Stat. Soc. Ser. B Stat. Methodol. 62 , (2000) , pp. 335--354. DOI:10.1111/1467-9868.00235.

J. Polzehl, V. Spokoiny, Propagation-separation approach for local likelihood estimation, Probab. Theory Related Fields 135 (3), (2006) , pp. 335--362. DOI:10.1007/s00440-005-0464-1.

Author

Joerg Polzehl, polzehl@wias-berlin.de, http://www.wias-berlin.de/people/polzehl/

Note

use setCores='number of threads' to enable parallel execution.

See also

Examples

if (FALSE) {
setCores(2)
y <- array(rnorm(4*64^3),c(4,64,64,64))
yhat <- vaws(y,kstar=20)
}